+++ apps/prairielearn/src/lib/externalGraderLocal.js
@@ -15,7 +15,7 @@ const { promisify } = require('util');
 
 const sql = sqldb.loadSqlEquiv(__filename);
 
-class Grader {
+export class ExternalGraderLocal {
   handleGradingRequest(grading_job, submission, variant, question, course) {
     const emitter = new EventEmitter();
 
@@ -23,7 +23,10 @@ class Grader {
 
     const dir = getDevJobDirectory(grading_job.id);
     const hostDir = getDevHostJobDirectory(grading_job.id);
-    const timeout = question.external_grading_timeout || config.externalGradingDefaultTimeout;
+    const timeout = Math.min(
+      question.external_grading_timeout ?? config.externalGradingDefaultTimeout,
+      config.externalGradingMaximumTimeout,
+    );
 
     const docker = new Docker();
 
@@ -238,5 +241,3 @@ function getDevHostJobDirectory(jobId) {
     return getDevJobDirectory(jobId);
   }
 }
-
-module.exports = Grader;