+++ apps/prairielearn/src/lib/question.sql
@@ -62,7 +62,6 @@ SELECT
   s.override_score,
   s.regradable,
   s.score,
-  s.sid,
   s.v2_score,
   s.variant_id,
   s.manual_rubric_grading_id,
@@ -240,10 +239,8 @@ WHERE
   s.id = $submission_id
   AND q.id = $question_id
   AND (
-    CASE
-      WHEN $instance_question_id::BIGINT IS NULL THEN iq.id IS NULL
-      ELSE iq.id = $instance_question_id::BIGINT
-    END
+    $instance_question_id::BIGINT IS NULL
+    OR iq.id = $instance_question_id::BIGINT
   )
   AND v.id = $variant_id;