+++ apps/prairielearn/src/models/course.sql
@@ -1,3 +1,11 @@
+-- BLOCK select_course_by_id
+SELECT
+  *
+FROM
+  pl_courses
+where
+  id = $course_id;
+
 -- BLOCK update_course_commit_hash
 UPDATE pl_courses
 SET