+++ packages/named-locks/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @prairielearn/named-locks
 
+## 1.5.8
+
+### Patch Changes
+
+- Updated dependencies [3d1c40c16]
+  - @prairielearn/postgres@1.8.0
+
+## 1.5.7
+
+### Patch Changes
+
+- 1523b97b0: Upgrade all dependencies
+- Updated dependencies [1523b97b0]
+  - @prairielearn/postgres@1.7.9
+
 ## 1.5.6
 
 ### Patch Changes