+++ packages/workspace-utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @prairielearn/workspace-utils
 
+## 1.0.14
+
+### Patch Changes
+
+- Updated dependencies [3d1c40c16]
+  - @prairielearn/postgres@1.8.0
+
+## 1.0.13
+
+### Patch Changes
+
+- Updated dependencies [1523b97b0]
+  - @prairielearn/postgres@1.7.9
+
 ## 1.0.12
 
 ### Patch Changes